+++ pih2e/ch01/Double.hs
@@ -1,0 +1,7 @@
+{-
+  file: ch01/Double.hs
+
+  This file isn't a Haskell module. You must load it inside `ghci`.
+-}
+
+double x = x + x